+++ CHANGES.rst
@@ -6,6 +6,8 @@
 Bug Fixes
 ---------
 
+- Add support for all valid astropy table write formats. [#3822]
+
 Cubeviz
 ^^^^^^^